+++ control.c
@@ -1,0 +1,8 @@
+/*
+  This file is part of EDWP
+  (Energy dispatcher work place)
+  Copyright (c) 2021 by Volodymyr Nerovnia
+  License: Apache 2.0
+*/
+
+#include "include/control.h"