+++ include/control.h
@@ -1,0 +1,12 @@
+/*
+  This file is part of EDWP
+  (Energy dispatcher work place)
+  Copyright (c) 2021 by Volodymyr Nerovnia
+  License: Apache 2.0
+*/
+
+#ifndef CONTROL_H_
+#define CONTROL_H_
+
+
+#endif /* CONTROL_H_ */