+++ include/form.h
@@ -1,0 +1,14 @@
+/*
+  This file is part of EDWP
+  (Energy dispatcher work place)
+  Copyright (c) 2021 by Volodymyr Nerovnia
+  License: Apache 2.0
+*/
+
+#ifndef FORM_H_
+#define FORM_H_
+
+#include <curses.h> 
+
+
+#endif /* FORM_H_ */