+++ test/convert.h
@@ -1,6 +1,0 @@
-#ifndef CONVERT_H_
-#define CONVERT_H_
-
-int conv_char_wchar_utf8(char*, wchar_t*);
-
-#endif /* CONVERT_H_ */